Refactor API endpoints in controllers for patient,appointments,insurance and billings

+++ HMS_Server/src/main/java/com/example/HMS/patients/controller/PatientController.java
@@ -31,6 +31,12 @@ public class PatientController {
         Patient createdPatient = patientService.createPatient(patient);
         return new ResponseEntity<>(createdPatient, HttpStatus.CREATED);
     }
+//    @PostMapping
+//    public ResponseEntity<List<Patient>> createPatients(@RequestBody List<Patient> patient) {
+//        List<Patient> createdPatients = patientService.createPatient(patient);
+//        return new ResponseEntity<>(createdPatients, HttpStatus.CREATED);
+//    }
+
 
     // Endpoint to get all patients
     @GetMapping
@@ -58,11 +64,21 @@ public class PatientController {
     }
 
     // Endpoint to delete a patient
+//    @DeleteMapping("/{id}")
+//    public ResponseEntity<Void> deletePatient(@PathVariable Long id) {
+//        patientService.deletePatient(id);
+//        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+//    }
+
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePatient(@PathVariable Long id) {
+        if (!patientService.existsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
         patientService.deletePatient(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
+
 }
